+++ Gobchat.App/Core/Util/DownloadHelper.cs
@@ -48,7 +48,7 @@ namespace Gobchat.Core.Util
             using (var webClient = new WebClient())
             {
                 var currentVersion = GobchatContext.ApplicationVersion;
-                webClient.Headers.Add("User-Agent", $"Gobchat v{currentVersion}");
+                webClient.Headers.Add("User-Agent", $"GobchatEx v{currentVersion}");
 
                 progressMonitor.Progress = 0d;
                 progressMonitor.StatusText = Resources.Core_Util_DownloadHelper_Waiting;